Use libusbg to configure the USB gadget
We need to have the gadget not configured for at least 700ms on
cable connect for bq24190 to detect the charger. So let's start
using libusbg to make it easier for us to enable and disable
the USB gadget.

Let's also compile it statically so we can use the same binary
for the installer initramfs.

Note that we now need libusbg compiled on the device to compile
this:

https://github.com/libusbg/libusbg

And we assume currently that libusbg directory is at ../libusbg.

And libusbg needs the following fix for charger detection
to actually clear the configured gadget:

http://marc.info/?l=linux-usb&m=141122826226028&w=2

Note that this means we can now remove the gadget configuration
from anvl-usb.sh.